rewrote button routing logic. added expandable button
