+++ app/login.js
@@ -10,7 +10,7 @@ import {
 	Image
 } from "react-native";
 import { styled } from "nativewind";
-import { Button } from "../components/Button";
+import { Button, ExpandableButton } from "../components/Buttons";
 import { Link } from "expo-router";
 import { loginUser } from "../database/firebaseFunctions";
 import AsyncStorage from "@react-native-async-storage/async-storage";
@@ -91,6 +91,7 @@ export default function Login() {
 								setPass("");
 							}}
 						/>
+						<ExpandableButton title="Profile"></ExpandableButton>
 						<StyledText className="text-offwhite text-center text-[18px] mt-5">
 							Don't have an account?{" "}
 							<Link href="/register">
